+++ app/resourses/service_resource.py
@@ -1,10 +1,9 @@
 from sanic.response import json, text
-from app.database import create_user, get_entry, delete_entry, insert_entry, update_entry, projects, invoices
+from marshmallow import ValidationError
+from app.database import create_user, get_entry, delete_entry, insert_entry,\
+    update_entry, projects, invoices
 from app.services.authorization import check_login_data
-
-
-async def smoke(request):
-    return text('hello, world!')
+from app.services.validation import ProjectsSchema, InvoicesSchema
 
 
 async def login(request):
@@ -19,8 +18,13 @@ async def project(request):
     if request.method == 'GET':
         return json(await get_entry(projects))
     if request.method == 'POST':
-        await insert_entry(projects, user_id=request.args.get('user_id'),
-                           create_date=request.args.get('create_date'))
+        json_data = request.args
+        try:
+            data = ProjectsSchema().load(json_data)
+        except ValidationError:
+            return json({'message': 'Invalid parameters'})
+        await insert_entry(projects, user_id=data[0]['user_id'],
+                           create_date=data[0]['create_date'])
     if request.method == 'DELETE':
         await delete_entry(projects)
         return text('projects deleted')
@@ -30,8 +34,13 @@ async def project_id(request, id_project):
     if request.method == 'GET':
         return json(await get_entry(projects, id_project))
     if request.method == 'PUT':
-        await update_entry(projects, id_project, user_id=request.args.get('user_id'),
-                           create_date=request.args.get('create_date'))
+        json_data = request.args
+        try:
+            data = ProjectsSchema().load(json_data)
+        except ValidationError:
+            return json({'message': 'Invalid parameters'})
+        await update_entry(projects, id_project, user_id=data[0]['user_id'],
+                           create_date=data[0]['create_date'])
     if request.method == 'DELETE':
         await delete_entry(projects, id_project)
         return text('projects deleted')
@@ -41,8 +50,13 @@ async def invoice(request, id_project):
     if request.method == 'GET':
         return json(await get_entry(invoices, project_id=id_project))
     if request.method == 'POST':
-        await insert_entry(invoices, project_id=request.args.get('project_id'),
-                           description=request.args.get('description'))
+        json_data = request.args
+        try:
+            data = InvoicesSchema().load(json_data)
+        except ValidationError:
+            return json({'message': 'Invalid parameters'})
+        await insert_entry(invoices, project_id=data[0]['project_id'],
+                           description=data[0]['description'])
     if request.method == 'DELETE':
         await delete_entry(invoices)
         return text('invoices deleted')
@@ -52,8 +66,13 @@ async def invoice_id(request, id_project, id_invoice):
     if request.method == 'GET':
         return json(await get_entry(invoices, id_project, id_invoice))
     if request.method == 'PUT':
-        await update_entry(invoices, id_invoice, project_id=request.args.get('project_id'),
-                           description=request.args.get('description'))
+        json_data = request.args
+        try:
+            data = InvoicesSchema().load(json_data)
+        except ValidationError:
+            return json({'message': 'Invalid parameters'})
+        await update_entry(invoices, id_invoice, project_id=data[0]['project_id'],
+                           description=data[0]['description'])
     if request.method == 'DELETE':
         await delete_entry(invoices, id_invoice)
         return text('invoices deleted')