+++ changelog.md
@@ -3,3 +3,4 @@
 * Setup Sanic server
 * Created endpoints for CRUD functions
 * Created class-storage for DB engine
+* Added validation for PUT and POST requests